various changes in merge
- use doubles for the weights
- in pureMkz, do not treat w=2 apart (for the moment)
- in pureMkz, new formula (to be checked)
- in MkzPopQueue, added check that mkz_count is correct (reveals errors)
- in merge_mono.c, replaced various constants 10000 by macro

+++ filter/filter_matrix.c
@@ -176,7 +176,7 @@ filter_matrix_read (filter_matrix_t *mat, purgedfile_stream_ptr ps, int verbose)
        Those columns should contribute to the average weight by
        sum(2/p, p prime < BURIED_MAX_DENSITY), which is ~ 4.58
        for BURIED_MAX_DENSITY=2000. */
-    wmax = mat->nrows / BURIED_MAX_DENSITY;
+    wmax = (int) ((double) mat->nrows / BURIED_MAX_DENSITY);
     /* heavy columns already have wt < 0 */
     tooheavy = (char *) malloc (mat->ncols * sizeof(char));
     memset (tooheavy, 0, mat->ncols * sizeof(char));
